Edit commit -- UI fixes. Made a called 'user_information'  page where users can update their personal information. None of the functionality on the 'user_information page' works yet. Deleted the multiform component and split each form component into individual components files.
