+++ src/app/globals.css
@@ -13,7 +13,7 @@ html {
   --background-end-rgb: 255, 255, 255;
   --color-primary: #f4ab9b;
   /* --color-primary:  #faebfa; */
-  --color-secondary: #b8f7d4;
+  --color-secondary: #ff9b9b;
   --color-primary-light: #fff8e3;
   --color-secondary-light: #ffdd95;
 }
