+++ src/utils/otherUtils.js
@@ -1,0 +1,16 @@
+import { object, string, date, ref } from "yup";
+
+const userSchema = object().shape({
+  firstName: string().required("Please type in your first name."),
+  lastName: string().required("Please type in your last name."),
+  date: date().required("Please type in your date of birth."),
+  gender: string().required("Please choose your gender. "),
+  email: string().email().required("Please type in your email."),
+  password: string().min(10).max(20).required("Please type in your password."),
+  confirmPassword: string().oneOf(
+    [ref("password"), null],
+    "Passwords must match"
+  ),
+});
+
+export default userSchema;
